Moved strategies to seperate file
