+++ lib/rom/setup/auto_registration_strategies/base.rb
@@ -1,0 +1,10 @@
+module ROM
+  module AutoRegistrationStrategies
+    class Base
+      EXTENSION_REGEX = /\.rb$/.freeze
+
+      include Options
+      option :file, reader: true, type: String
+    end
+  end
+end
