+++ lib/rom/setup/auto_registration_strategies/with_namespace.rb
@@ -1,0 +1,13 @@
+module ROM
+  module AutoRegistrationStrategies
+    class WithNamespace < Base
+      option :directory, reader: true, type: Pathname
+
+      def call
+        Inflector.camelize(
+          file.sub(/^#{directory.dirname}\//, '').sub(EXTENSION_REGEX, '')
+        )
+      end
+    end
+  end
+end
